+++ wp-admin/includes/nav-menu.php
@@ -874,7 +874,7 @@ function wp_nav_menu_item_taxonomy_meta_box( $data_object, $box ) {
 		return;
 	}
 
-	$num_pages = ceil(
+	$num_pages = (int) ceil(
 		wp_count_terms(
 			array_merge(
 				$args,
