+++ wp-includes/class-wp-site-query.php
@@ -377,7 +377,7 @@ class WP_Site_Query {
 		}
 
 		if ( $this->found_sites && $this->query_vars['number'] ) {
-			$this->max_num_pages = ceil( $this->found_sites / $this->query_vars['number'] );
+			$this->max_num_pages = (int) ceil( $this->found_sites / $this->query_vars['number'] );
 		}
 
 		// If querying for a count only, there's nothing more to do.
