+++ wp-includes/comment-template.php
@@ -1515,7 +1515,7 @@ function comments_template( $file = '/comments.php', $separate_comments = false 
 
 			$top_level_count = $top_level_query->query( $top_level_args );
 
-			$comment_args['offset'] = ( ceil( $top_level_count / $per_page ) - 1 ) * $per_page;
+			$comment_args['offset'] = ( (int) ceil( $top_level_count / $per_page ) - 1 ) * $per_page;
 		}
 	}
